Capstone Inclusion
White-Out is a collaborative notation project designed for use in conjunction with a screen sharing service. White-Out initially was built with the use of an AWS EC2 instance as a primary server, but has since been redesigned to allow for any two machines to act as client and server as long as the server's IP address is visible to the clients.